refactor: separate DNS service with dedicated router and adapter
