+++ app/api/__init__.py
@@ -9,8 +9,8 @@ from .auth.router_auth import auth_router
 from .auth.router_mfa import mfa_router
 from .auth.session_router import session_router
 from .dhcp.router import dhcp_router
+from .dns.router import dns_router
 from .ldap_schema.entity_type_router import ldap_schema_router
-from .main.dns_router import dns_router
 from .main.krb5_router import krb5_router
 from .main.router import entry_router
 from .network.router import network_router
