+++ app/api/dns/adapter.py
